requirements.txt
albucore==0.0.24
albumentations==2.0.8
alembic==1.17.2
annotated-types==0.7.0
asttokens==3.0.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.1
debugpy==1.8.17
decorator==5.2.1
easydict==1.13
executing==2.2.1
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
greenlet==3.2.4
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
insightface==0.7.3
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
ml_dtypes==0.5.3
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
onnx==1.19.1
onnxruntime-gpu==1.18.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.0
polars==1.35.2
polars-runtime-32==1.35.2
prettytable==3.16.0
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
psycopg2-binary==2.9.11
pure_eval==0.2.3
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
python-multipart==0.0.9
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
simsimd==6.5.3
six==1.17.0
SQLAlchemy==2.0.44
stack-data==0.6.3
stringzilla==4.2.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.10.16
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.228
ultralytics-thop==2.0.18
urllib3==2.5.0
wcwidth==0.2.14
